Make SIMD module a submodule of firstpass
This makes it clear that the SIMD module is only used to accelerate
the first pass.
